requirements.txt
aiohttp
pyyaml
aiopg[sa]
aiohttp_jinja2
flake8
pytest
pytest-aiohttp
